refactor(frontend): standardize file naming to Angular 21 convention (#260)
Per Angular 20+ style guide, remove .component suffix from file names:
- user-profile.component.ts -> user-profile.ts
- user-profile.component.html -> user-profile.html
- user-profile.component.css -> user-profile.css

Changes:
- Rename 43 component files (auth, components folders)
- Update templateUrl/styleUrl references
- Update all imports throughout codebase
- Remove unused placeholder files (task-create)

This aligns with Angular's updated style guide which recommends
simpler file names for standalone components.

Closes #260
