+++ apps/frontend/src/app/auth/reset-password.ts
@@ -11,8 +11,8 @@ import {
 
 @Component({
   selector: 'app-reset-password',
-  templateUrl: './reset-password.component.html',
-  styleUrl: './reset-password.component.css',
+  templateUrl: './reset-password.html',
+  styleUrl: './reset-password.css',
   imports: [CommonModule, ReactiveFormsModule, RouterLink],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
