+++ apps/frontend/src/app/components/manual-assignment-modal/manual-assignment-modal.html
