+++ apps/frontend/src/app/components/task-card/task-card.css
