--- apps/frontend/src/app/components/task-create/task-create.component.css
@@ -1,256 +1,0 @@
-.task-create-form {
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-h2 {
-  margin-bottom: 1.5rem;
-  font-size: 1.5rem;
-  font-weight: 600;
-}
-
-.form-group {
-  margin-bottom: 1.5rem;
-}
-
-label,
-legend {
-  display: block;
-  margin-bottom: 0.5rem;
-  font-weight: 500;
-  color: #374151;
-}
-
-.required {
-  color: #dc2626;
-}
-
-.help {
-  font-size: 0.875rem;
-  font-weight: 400;
-  color: #6b7280;
-}
-
-input[type='text'],
-textarea {
-  width: 100%;
-  padding: 0.5rem;
-  border: 1px solid #d1d5db;
-  border-radius: 0.375rem;
-  font-size: 1rem;
-}
-
-input[type='text']:focus,
-textarea:focus {
-  outline: none;
-  border-color: #3b82f6;
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
-}
-
-input[type='text'][aria-invalid='true'] {
-  border-color: #dc2626;
-}
-
-.char-count {
-  float: right;
-  font-size: 0.875rem;
-  color: #6b7280;
-  margin-top: 0.25rem;
-}
-
-fieldset {
-  border: none;
-  padding: 0;
-  margin: 0;
-}
-
-.rule-types,
-.rotation-types {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-
-.rule-option,
-.rotation-option {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.75rem;
-  padding: 1rem;
-  border: 2px solid #e5e7eb;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.rule-option:hover,
-.rotation-option:hover {
-  border-color: #3b82f6;
-  background-color: #f0f9ff;
-}
-
-.rule-option:has(input:checked),
-.rotation-option:has(input:checked) {
-  border-color: #3b82f6;
-  background-color: #dbeafe;
-}
-
-.rule-option input[type='radio'],
-.rotation-option input[type='radio'] {
-  margin-top: 0.25rem;
-  cursor: pointer;
-}
-
-.rule-option div,
-.rotation-option div {
-  flex: 1;
-}
-
-.rule-option strong,
-.rotation-option strong {
-  display: block;
-  margin-bottom: 0.25rem;
-  color: #111827;
-}
-
-.rule-option p,
-.rotation-option p {
-  margin: 0;
-  font-size: 0.875rem;
-  color: #6b7280;
-}
-
-.day-selector,
-.children-selector {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.75rem;
-}
-
-.day-checkbox,
-.child-checkbox {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.5rem 0.75rem;
-  border: 1px solid #d1d5db;
-  border-radius: 0.375rem;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.day-checkbox:hover,
-.child-checkbox:hover {
-  border-color: #3b82f6;
-  background-color: #f0f9ff;
-}
-
-.day-checkbox:has(input:checked),
-.child-checkbox:has(input:checked) {
-  border-color: #3b82f6;
-  background-color: #dbeafe;
-  font-weight: 500;
-}
-
-.day-checkbox input[type='checkbox'],
-.child-checkbox input[type='checkbox'] {
-  cursor: pointer;
-}
-
-.error {
-  margin-top: 0.5rem;
-  font-size: 0.875rem;
-  color: #dc2626;
-}
-
-.success-message {
-  padding: 1rem;
-  margin-bottom: 1rem;
-  background-color: #dcfce7;
-  border: 1px solid #86efac;
-  border-radius: 0.375rem;
-  color: #166534;
-}
-
-.error-message {
-  padding: 1rem;
-  margin-bottom: 1rem;
-  background-color: #fee2e2;
-  border: 1px solid #fca5a5;
-  border-radius: 0.375rem;
-  color: #991b1b;
-}
-
-.form-actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 1rem;
-  margin-top: 2rem;
-}
-
-button {
-  padding: 0.5rem 1.5rem;
-  font-size: 1rem;
-  font-weight: 500;
-  border-radius: 0.375rem;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-button:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.btn-secondary {
-  background-color: white;
-  color: #374151;
-  border: 1px solid #d1d5db;
-}
-
-.btn-secondary:hover:not(:disabled) {
-  background-color: #f9fafb;
-}
-
-.btn-primary {
-  background-color: #3b82f6;
-  color: white;
-  border: none;
-}
-
-.btn-primary:hover:not(:disabled) {
-  background-color: #2563eb;
-}
-
-.btn-primary:focus,
-.btn-secondary:focus {
-  outline: none;
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.3);
-}
-
-/* Responsive */
-@media (max-width: 640px) {
-  .task-create-form {
-    padding: 1rem;
-  }
-
-  .day-selector,
-  .children-selector {
-    gap: 0.5rem;
-  }
-
-  .day-checkbox,
-  .child-checkbox {
-    padding: 0.375rem 0.5rem;
-    font-size: 0.875rem;
-  }
-
-  .form-actions {
-    flex-direction: column-reverse;
-  }
-
-  button {
-    width: 100%;
-  }
-}
