+++ apps/frontend/src/app/components/task-create/task-create.html
@@ -1,1 +1,217 @@
-<p>task-create works!</p>
+<div class="task-create-form">
+  <h2>Create Task Template</h2>
+
+  <form [formGroup]="taskForm" (ngSubmit)="onSubmit()">
+    <!-- Name -->
+    <div class="form-group">
+      <label for="name"> Task Title <span class="required">*</span> </label>
+      <input
+        id="name"
+        type="text"
+        formControlName="name"
+        placeholder="e.g., Take out trash"
+        [attr.aria-invalid]="taskForm.get('name')?.invalid && taskForm.get('name')?.touched"
+        [attr.aria-describedby]="
+          taskForm.get('name')?.invalid && taskForm.get('name')?.touched ? 'name-error' : null
+        "
+      />
+      <span class="char-count" aria-live="polite">{{ nameChars }}/200</span>
+      @if (taskForm.get('name')?.invalid && taskForm.get('name')?.touched) {
+        <p id="name-error" class="error" role="alert">Title is required (max 200 characters)</p>
+      }
+    </div>
+
+    <!-- Description -->
+    <div class="form-group">
+      <label for="description">Description (Optional)</label>
+      <textarea
+        id="description"
+        formControlName="description"
+        rows="3"
+        placeholder="e.g., Empty all trash bins, including bathroom"
+      ></textarea>
+    </div>
+
+    <!-- Rule Type -->
+    <div class="form-group">
+      <fieldset>
+        <legend>Rule Type <span class="required">*</span></legend>
+        <div class="rule-types">
+          <label class="rule-option">
+            <input
+              type="radio"
+              formControlName="ruleType"
+              value="daily"
+              aria-describedby="daily-desc"
+            />
+            <div>
+              <strong>Daily</strong>
+              <p id="daily-desc">Task assigned every day</p>
+            </div>
+          </label>
+          <label class="rule-option">
+            <input
+              type="radio"
+              formControlName="ruleType"
+              value="repeating"
+              aria-describedby="repeating-desc"
+            />
+            <div>
+              <strong>Repeating</strong>
+              <p id="repeating-desc">Task on specific days of the week</p>
+            </div>
+          </label>
+          <label class="rule-option">
+            <input
+              type="radio"
+              formControlName="ruleType"
+              value="weekly_rotation"
+              aria-describedby="rotation-desc"
+            />
+            <div>
+              <strong>Weekly Rotation</strong>
+              <p id="rotation-desc">Alternates between children each week</p>
+            </div>
+          </label>
+        </div>
+      </fieldset>
+    </div>
+
+    <!-- Repeat Days (shown if repeating) -->
+    @if (ruleType === 'repeating') {
+      <div class="form-group">
+        <fieldset>
+          <legend>Repeat On <span class="required">*</span></legend>
+          <div class="day-selector" role="group" aria-label="Select days">
+            @for (day of daysOfWeek; track day.value) {
+              <label class="day-checkbox">
+                <input
+                  type="checkbox"
+                  [checked]="isDaySelected(day.value)"
+                  (change)="onDayChange(day.value, $any($event.target).checked)"
+                  [attr.aria-label]="'Repeat on ' + day.label"
+                />
+                {{ day.label }}
+              </label>
+            }
+          </div>
+          @if (taskForm.get('repeatDays')?.invalid && taskForm.get('repeatDays')?.touched) {
+            <p class="error" role="alert">Select at least one day</p>
+          }
+        </fieldset>
+      </div>
+    }
+
+    <!-- Rotation Type (shown if weekly_rotation) -->
+    @if (ruleType === 'weekly_rotation') {
+      <div class="form-group">
+        <fieldset>
+          <legend>Rotation Type <span class="required">*</span></legend>
+          <div class="rotation-types">
+            <label class="rotation-option">
+              <input
+                type="radio"
+                formControlName="rotationType"
+                value="odd_even_week"
+                aria-describedby="odd-even-desc"
+              />
+              <div>
+                <strong>Odd/Even Week</strong>
+                <p id="odd-even-desc">Different child each calendar week</p>
+              </div>
+            </label>
+            <label class="rotation-option">
+              <input
+                type="radio"
+                formControlName="rotationType"
+                value="alternating"
+                aria-describedby="alternating-desc"
+              />
+              <div>
+                <strong>Alternating</strong>
+                <p id="alternating-desc">Rotates every 7 days from start</p>
+              </div>
+            </label>
+          </div>
+          @if (taskForm.get('rotationType')?.invalid && taskForm.get('rotationType')?.touched) {
+            <p class="error" role="alert">Please select a rotation type</p>
+          }
+        </fieldset>
+      </div>
+    }
+
+    <!-- Assigned Children -->
+    @if (ruleType !== 'daily') {
+      <div class="form-group">
+        <fieldset>
+          <legend>
+            Assign to Children <span class="required">*</span>
+            @if (ruleType === 'weekly_rotation') {
+              <span class="help">(Select 2+ for rotation)</span>
+            }
+          </legend>
+          <div class="children-selector" role="group" aria-label="Assign children">
+            @for (child of children(); track child.id) {
+              <label class="child-checkbox">
+                <input
+                  type="checkbox"
+                  [checked]="isChildSelected(child.id)"
+                  (change)="onChildChange(child.id, $any($event.target).checked)"
+                  [attr.aria-label]="'Assign to ' + child.name"
+                />
+                {{ child.name }}
+              </label>
+            }
+          </div>
+          @if (
+            taskForm.get('assignedChildren')?.invalid && taskForm.get('assignedChildren')?.touched
+          ) {
+            <p class="error" role="alert">
+              @if (ruleType === 'weekly_rotation') {
+                Select at least 2 children for rotation
+              } @else {
+                Select at least 1 child
+              }
+            </p>
+          }
+        </fieldset>
+      </div>
+    }
+
+    <!-- Messages -->
+    @if (successMessage()) {
+      <div class="success-message" role="status" aria-live="polite">
+        {{ successMessage() }}
+      </div>
+    }
+    @if (taskServiceError) {
+      <div class="error-message" role="alert" aria-live="assertive">
+        {{ taskServiceError }}
+      </div>
+    }
+
+    <!-- Actions -->
+    <div class="form-actions">
+      <button
+        type="button"
+        (click)="onCancel()"
+        class="btn-secondary"
+        [disabled]="taskServiceLoading"
+      >
+        Cancel
+      </button>
+      <button
+        type="submit"
+        class="btn-primary"
+        [disabled]="taskForm.invalid || taskServiceLoading"
+        [attr.aria-busy]="taskServiceLoading"
+      >
+        @if (taskServiceLoading) {
+          Creating...
+        } @else {
+          Create Task
+        }
+      </button>
+    </div>
+  </form>
+</div>
